Move benchmarks to their own test files 🏡
